requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
django==5.1.7
gunicorn==21.2.0
hiredis==3.1.0
jinja2==3.1.6
kombu==5.5.1
markupsafe==3.0.2
prompt-toolkit==3.0.50
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
supervisor==4.2.5
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
